Moved EventDataDecorators to new package
