+++ collector-bitmovin-player/src/test/java/com/bitmovin/analytics/bitmovin/player/BitmovinSdkAdapterTest.kt
@@ -1,7 +1,7 @@
 package com.bitmovin.analytics.bitmovin.player
 
 import com.bitmovin.analytics.BitmovinAnalyticsConfig
-import com.bitmovin.analytics.data.DeviceInformationEventDataDecorator
+import com.bitmovin.analytics.data.decorators.DeviceInformationEventDataDecorator
 import com.bitmovin.analytics.stateMachines.PlayerState
 import com.bitmovin.analytics.stateMachines.PlayerStateMachine
 import com.bitmovin.player.BitmovinPlayer
