+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/ExoPlayerAdapter.java
@@ -9,7 +9,7 @@ import com.bitmovin.analytics.adapters.PlayerAdapter;
 import com.bitmovin.analytics.data.DRMInformation;
 import com.bitmovin.analytics.data.ErrorCode;
 import com.bitmovin.analytics.data.EventData;
-import com.bitmovin.analytics.data.DeviceInformationEventDataDecorator;
+import com.bitmovin.analytics.data.decorators.DeviceInformationEventDataDecorator;
 import com.bitmovin.analytics.data.EventDataDecorator;
 import com.bitmovin.analytics.data.SpeedMeasurement;
 import com.bitmovin.analytics.enums.DRMType;
