+++ collector-exoplayer/src/test/java/com/bitmovin/analytics/exoplayer/FakeExoPlayerAdapter.kt
@@ -2,7 +2,7 @@ package com.bitmovin.analytics.exoplayer
 
 import com.bitmovin.analytics.BitmovinAnalyticsConfig
 import com.bitmovin.analytics.EventDataDecoratorPipeline
-import com.bitmovin.analytics.data.DeviceInformationEventDataDecorator
+import com.bitmovin.analytics.data.decorators.DeviceInformationEventDataDecorator
 import com.bitmovin.analytics.stateMachines.PlayerStateMachine
 import com.google.android.exoplayer2.ExoPlayer
