+++ collector/src/test/java/com/bitmovin/analytics/data/decorators/DeviceInformationEventDataDecoratorTest.kt
@@ -1,0 +1,5 @@
+package com.bitmovin.analytics.data.decorators
+
+import junit.framework.TestCase
+
+class DeviceInformationEventDataDecoratorTest : TestCase()
